mcp/pyproject.toml
[project]
name = "word-mcp"
version = "0.1.0"
description = "MCP server for editing Word documents with Track Changes"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0",
    "python-docx>=1.1",
    "lxml>=5.0",
    "msal>=1.20",
    "httpx>=0.27",
    "starlette>=0.40",
]

[project.scripts]
word-mcp = "word_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/word_mcp"]
